Fixed sign conversion error and unused includes

+++ program/source/window.cpp
@@ -1,3 +1,4 @@
+#include <cstdint>
 #include <iostream>
 
 #include "rename.hpp"
@@ -108,8 +109,8 @@ namespace cse::system
 #ifdef _WIN32
     sdl::set_window_bordered(application, SDL_FALSE);
     sdl::set_window_size(application, display_mode.w, display_mode.h + 1);
-    sdl::set_window_position(application, SDL_WINDOWPOS_CENTERED_DISPLAY(display_index),
-                             SDL_WINDOWPOS_CENTERED_DISPLAY(display_index));
+    sdl::set_window_position(application, (int)SDL_WINDOWPOS_CENTERED_DISPLAY(display_index),
+                             (int)SDL_WINDOWPOS_CENTERED_DISPLAY(display_index));
 #endif
 #ifdef __linux__
     sdl::set_window_fullscreen(application, SDL_WINDOW_FULLSCREEN_DESKTOP);
